In the 'getTasks' function, the data is initially retrieved from the local Room database. If the data is not empty (meaning it has been previously fetched from the API and added to the database), it retrieves the data from there. If the data is being fetched for the first time (i.e., when the user logs in for the first time), the 'else' part in 'getTasks' is triggered, attempting to fetch data from the API. If the data can be fetched, it's added to the database; otherwise, an error is returned. If the error is an authorization issue, the application automatically redirects the user to the login screen.
In 'updateTasks', it first tries to fetch data from the API. If the fetching is successful, all previous data in the database is cleared, and the new data from the API is added to the database. Then, the data is retrieved from the database and presented to the user. If fetching data from the API fails but the local Room database is not empty, the data is fetched from Room, and an error message, such as 'Data couldn't be fetched from the API,' is shown to the user. If fetching from the API fails and the local database is empty, an error message is displayed again. In both error scenarios, if there's an authorization issue, the user is redirected to the login screen.